move get params to C

+++ js/communication.js
@@ -160,7 +160,7 @@ ws.onmessage = function(event) {
             // it might be worth it to have some of the things preallocated and cached.
             // 1. Figure out if it is a bottleneck. When things are chunked it should be easier to reuse
             spectrogram.updateStartLoadTime();
-            spectrogram.updateSpectrogram(new Float32Array(event.data, headerLen + 4),
+            spectrogram.updateSpectrogram(new Float64Array(event.data, headerLen + 4),
                 content.nblocks, content.nfreqs);
             spectrogram.logElaspedTime();
         } else {
